feat:add solution for fs.delete, and refactoring pathFiles to fs.copy and fs.create
