+++ src/fs/create.js
@@ -1,16 +1,22 @@
-import fs from "fs";
+import fs from "fs/promises";
+import url from "url";
+import path from "path";
 const create = async () => {
-  fs.exists("./files/fresh.txt", (exist) => {
-    if (exist) {
-      throw new Error("FS operation failed");
-    } else {
-      fs.writeFile("./files/fresh.txt", "I am fresh and young", (error) => {
-        if (error) {
-          throw new Error("FS operation failed");
-        }
-      });
-    }
-  });
+  const currentDir = path.dirname(url.fileURLToPath(import.meta.url));
+  const searchFilePath = path.join(currentDir, "files", "fresh.txt");
+  const searchFile = await fs
+    .stat(searchFilePath)
+    .then(() => true)
+    .catch(() => false);
+  if (!searchFile) {
+    fs.writeFile(searchFilePath, "I am fresh and young", (error) => {
+      if (error) {
+        console.log(error);
+      }
+    });
+  } else {
+    throw new Error("FS operation failed");
+  }
 };
 
 await create();
